Tracking: Split submodule out into separate files
Split the main module into several, more focussed files.
