+++ kano_profile/tracker/__init__.py
@@ -4,12 +4,15 @@
 # Copyright (C) 2014 - 2017 Kano Computing Ltd.
 # License: http://www.gnu.org/licenses/gpl-2.0.txt GNU GPL v2
 #
+# Kano-tracker module
+#
+# A small module for tracking various metrics the users do in Kano OS
+#
 
-"""
-Kano-tracker module
 
-A small module for tracking various metrics the users do in Kano OS
-"""
+__author__ = 'Kano Computing Ltd.'
+__email__ = 'dev@kano.me'
+
 
 import time
 import atexit
@@ -23,204 +26,27 @@ import shlex
 
 from uuid import uuid1, uuid5
 
-from kano.utils import get_program_name, is_number, read_file_contents, \
-    get_cpu_id, chown_path, ensure_dir
+from kano.utils.processes import get_program_name
+from kano.utils.file_operations import read_file_contents, chown_path, \
+    ensure_dir
+from kano.utils.hardware import get_cpu_id
+from kano.utils.misc import is_number
 from kano.logging import logger
 from kano_profile.apps import get_app_state_file, load_app_state_variable, \
     save_app_state_variable
 from kano_profile.paths import tracker_dir, tracker_events_file, \
     tracker_token_file
 
+from kano_profile.tracker.tracker_token import TOKEN, generate_tracker_token, \
+       load_token
+from kano_profile.tracker.tracking_utils import open_locked, \
+    _get_nearest_previous_monday, get_utc_offset
 
-class open_locked(file):
-    """ A version of open with an exclusive lock to be used within
-        controlled execution statements.
-    """
-    def __init__(self, *args, **kwargs):
-        super(open_locked, self).__init__(*args, **kwargs)
-        fcntl.flock(self, fcntl.LOCK_EX)
-
-
-def load_token():
-    """
-        Reads the tracker token from the token file. If the file doesn't
-        exists, it regenerates it.
-
-        The token is regenerated on each boot and is inserted into every
-        event to link together events that happened during the same start
-        of the OS.
-
-        :returns: The token.
-        :rtype: str
-    """
-
-    if os.path.exists(tracker_token_file):
-        try:
-            f = open_locked(tracker_token_file, 'r')
-        except IOError as e:
-            logger.error("Error opening tracker token file {}".format(e))
-        else:
-            with f:
-                return f.read().strip()
-    else:
-        return generate_tracker_token()
-
-
-def generate_tracker_token():
-    """
-        Generating the token is a simple md5hash of the current time.
-
-        The token is saved to the `tracker_token_file`.
-
-        :returns: The token.
-        :rtype: str
-    """
-
-    token = hashlib.md5(str(time.time())).hexdigest()
-
-    ensure_dir(tracker_dir)
-    try:
-        f = open_locked(tracker_token_file, 'w')
-    except IOError as e:
-        logger.error(
-            "Error opening tracker token file (generate) {}".format(e))
-    else:
-        with f:
-            f.write(token)
-        if 'SUDO_USER' in os.environ:
-            chown_path(tracker_token_file)
-
-    # Make sure that the events file exist
-    try:
-        f = open(tracker_events_file, 'a')
-    except IOError as e:
-        logger.error("Error opening tracker events file {}".format(e))
-    else:
-        f.close()
-        if 'SUDO_USER' in os.environ:
-            chown_path(tracker_events_file)
-
-    return token
-
-
-OS_VERSION = str(read_file_contents('/etc/kanux_version'))
-os_variant = read_file_contents('/etc/kanux_version_variant')
-OS_VERSION += '-' + os_variant if os_variant else ''
-CPU_ID = str(get_cpu_id())
-TOKEN = load_token()
-LANGUAGE = (os.getenv('LANG') or '').split('.', 1)[0]
-
-
-def get_session_file_path(name, pid):
-    return "{}/{}-{}.json".format(tracker_dir, pid, name)
-
-
-def get_session_unique_id(name, pid):
-    data = {}
-    tracker_session_file = get_session_file_path(name, pid)
-    try:
-        af = open_locked(tracker_session_file, 'r')
-    except (IOError, OSError) as e:
-        logger.error("Error while opening session file".format(e))
-    else:
-        with af:
-            try:
-                data = json.load(af)
-            except ValueError as e:
-                logger.error("Session file is not a valid JSON")
-
-    return data.get('app_session_id', "")
-
-
-def session_start(name, pid=None):
-    if not pid:
-        pid = os.getpid()
-    pid = int(pid)
-
-    data = {
-        'pid': pid,
-        'name': name,
-        'started': int(time.time()),
-        'elapsed': 0,
-        'app_session_id': str(uuid5(uuid1(), name + str(pid))),
-        'finished': False,
-        'token-system': TOKEN
-    }
-
-    path = get_session_file_path(data['name'], data['pid'])
-
-    try:
-        f = open_locked(path, 'w')
-    except IOError as e:
-        logger.error("Error opening tracker session file {}".format(e))
-    else:
-        with f:
-            json.dump(data, f)
-        if 'SUDO_USER' in os.environ:
-            chown_path(path)
-
-    return path
-
-
-def session_end(session_file):
-    if not os.path.exists(session_file):
-        msg = "Someone removed the tracker file, the runtime of this " \
-              "app will not be logged"
-        logger.warn(msg)
-        return
-
-    try:
-        rf = open_locked(session_file, 'r')
-    except IOError as e:
-        logger.error("Error opening the tracker session file {}".format(e))
-    else:
-        with rf:
-            data = json.load(rf)
-
-            data['elapsed'] = int(time.time()) - data['started']
-            data['finished'] = True
-
-            try:
-                wf = open(session_file, 'w')
-            except IOError as e:
-                logger.error(
-                    "Error opening the tracker session file {}".format(e))
-            else:
-                with wf:
-                    json.dump(data, wf)
-        if 'SUDO_USER' in os.environ:
-            chown_path(session_file)
-
-
-def session_log(name, started, length):
-    """ Log a session that was tracked outside of the tracker.
-
-        :param name: The identifier of the session.
-        :type name: str
-
-        :param started: When was the session started (UTC unix timestamp).
-        :type started: int
-
-        :param length: Length of the session in seconds.
-        :param started: int
-    """
-
-    try:
-        af = open_locked(tracker_events_file, 'a')
-    except IOError as e:
-        logger.error("Error while opening events file".format(e))
-    else:
-        with af:
-            session = {
-                'name': name,
-                'started': int(started),
-                'elapsed': int(length)
-            }
-
-            event = get_session_event(session)
-            af.write(json.dumps(event) + "\n")
-        if 'SUDO_USER' in os.environ:
-            chown_path(tracker_events_file)
+# Public imports
+from kano_profile.tracker.tracker import Tracker
+from kano_profile.tracker.tracking_sessions import session_start, session_end, \
+    get_session_file_path, session_log, get_session_unique_id, \
+    get_session_event, CPU_ID, LANGUAGE, OS_VERSION
 
 
 def track_data(name, data):
@@ -306,62 +132,6 @@ def get_action_event(name):
         'language': LANGUAGE,
         'name': name
     }
-
-
-def get_session_event(session):
-    """ Construct the event data structure for a session. """
-
-    return {
-        'type': 'session',
-        'time': session['started'],
-        'timezone_offset': get_utc_offset(),
-        'os_version': OS_VERSION,
-        'cpu_id': CPU_ID,
-        'token': TOKEN,
-        'language': LANGUAGE,
-        'name': session['name'],
-        'length': session['elapsed'],
-        'token-system': session.get('token-system', '')
-    }
-
-
-def get_utc_offset():
-    """ Returns the local UTC offset in seconds.
-
-        :returns: UTC offsed in secconds.
-        :rtype: int
-    """
-
-    is_dst = time.daylight and time.localtime().tm_isdst > 0
-    return -int(time.altzone if is_dst else time.timezone)
-
-
-class Tracker:
-    """Tracker class, used for measuring program run-times,
-    implemented via atexit hooks"""
-
-    def __init__(self):
-        self.path = session_start(get_program_name())
-        atexit.register(self._write_times)
-
-    def _write_times(self):
-        session_end(self.path)
-
-
-# TODO: While it isn't at the moment, this could be useful to have
-#       in the toolset.
-def _get_nearest_previous_monday():
-    """ Returns the timestamp of the nearest past Monday """
-
-    t = time.time()
-    day = 24 * 60 * 60
-    week = 7 * day
-
-    r = (t - (t % week)) - (3 * day)
-    if (t - r) >= week:
-        r += week
-
-    return int(r)
 
 
 def add_runtime_to_app(app, runtime):
